requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-import-export==3.3.1
whitenoise==6.5.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
gunicorn==21.2.0
django-jazzmin 
xhtml2pdf
